requirements.txt
requests
PyYAML
websockets
pycryptodome
google-generativeai